refactor: clean up clippy warnings across workspace
Fix all clippy warnings to achieve zero-warning build:
- Replace from_str() with parse() to avoid FromStr trait confusion
- Collapse nested if statements, simplify map_or to is_some_and
- Extract type aliases for complex return types
- Add #[allow] for intentional too_many_arguments/await_holding_lock
- Remove dead code placeholders and unused imports
- Fix manual clamp pattern, collapsible if, owned instance comparison

+++ crates/agent/src/runtime/decision/http/cognitive_context.rs
@@ -316,7 +316,7 @@ impl CognitiveContextBuilder {
             for (trait_name, trait_obj) in p.traits.iter().take(2) {
                 drives.push(Drive {
                     drive: trait_name.clone(),
-                    intensity: (trait_obj.value() as u8 / 2).max(3),
+                    intensity: (trait_obj.value() / 2).max(3),
                     reason: format!("基于{}的性格倾向", p.name),
                 });
             }
